Added New melee enemy type and revised Enemy prefabs

+++ Puzzle FPS/Assets/Scripts/EnemyBullet.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class EnemyBullet : MonoBehaviour
 {
     public int bulletDamage;
-    [SerializeField] int timer;
+    [SerializeField] float timer;
 
     // Start is called before the first frame update
     void Start()
